Rename: AuthUser 패키지 복구
