+++ src/main/java/com/ormi/mogakcote/comment/presentation/CommentController.java
@@ -1,6 +1,6 @@
 package com.ormi.mogakcote.comment.presentation;
 
-import com.ormi.mogakcote.security.model.AuthUser;
+import com.ormi.mogakcote.auth.model.AuthUser;
 import com.ormi.mogakcote.comment.application.CommentService;
 import com.ormi.mogakcote.comment.dto.request.CommentRequest;
 import com.ormi.mogakcote.comment.dto.response.CommentResponse;
